Sanity checks on uploaded screenshots.

+++ trunk/shotserver/lib/xmlrpc/request.py
@@ -20,10 +20,11 @@ __revision__ = '$Rev$'
 __date__ = '$Date$'
 __author__ = '$Author$'
 
-import os, tempfile
+import os, tempfile, re
 from shotserver03 import database
 
 export_methods = ['poll', 'upload']
+pngpath = '/var/www/browsershots.org/png'
 
 def poll(factory, crypt):
     """
@@ -81,10 +82,64 @@ def poll(factory, crypt):
     finally:
         database.disconnect()
 
-def zoom(ppmname, pngpath, prefix, hashkey, width):
+header_match = re.compile(r'(P\d) (\d+) (\d+) (\d+)').match
+def read_ppm_header(infile):
+    """
+    Read a PPM file header and return magic, width, height, maxval.
+    """
+    header = []
+    while True:
+        line = infile.readline()
+        sharp = line.find('#')
+        if sharp > -1:
+            line = line[:sharp]
+        line = line.strip()
+        if not line:
+            continue
+        header.append(line)
+        match = header_match(' '.join(header))
+        if match:
+            magic = match.group(1)
+            width = int(match.group(2))
+            height = int(match.group(3))
+            maxval = int(match.group(4))
+            return magic, width, height, maxval
+        elif len(header) >= 4:
+            raise SyntaxError("could not parse PPM header")
+
+def save_upload(binary, hashkey):
+    """
+    Save the upload to a PNG file.
+    """
+    prefix = hashkey[:2]
+    fullpath = '%s/full/%s' % (pngpath, prefix)
+    if not os.path.exists(fullpath):
+        os.makedirs(fullpath)
+    pngname = '%s/%s.png' % (fullpath, hashkey)
+    outfile = file(pngname, 'wb')
+    outfile.write(binary.data)
+    outfile.close()
+
+def pngtoppm(hashkey):
+    """
+    Decode the uploaded PNG file to a temporary PPM.
+    """
+    prefix = hashkey[:2]
+    fullpath = '%s/full/%s' % (pngpath, prefix)
+    pngname = '%s/%s.png' % (fullpath, hashkey)
+    ppmhandle, ppmname = tempfile.mkstemp()
+    error = os.system('pngtopnm "%s" > "%s"' % (pngname, ppmname))
+    assert not error
+    magic, width, height, maxval = read_ppm_header(file(ppmname))
+    assert magic == 'P6'
+    assert maxval == 255
+    return width, height, ppmhandle, ppmname
+
+def zoom(ppmname, hashkey, width):
     """
     Make smaller preview images.
     """
+    prefix = hashkey[:2]
     zoompath = '%s/%d/%s' % (pngpath, width, prefix)
     if not os.path.exists(zoompath):
         os.makedirs(zoompath)
@@ -97,31 +152,32 @@ def upload(binary, crypt):
     """
     Upload a browser screenshot.
     """
-    pngpath = '/var/www/browsershots.org/png'
     database.connect()
     try:
         ip = req.connection.remote_ip
-        status, request = database.nonce.authenticate_request(ip, crypt)
+        status, request, request_width, factory = database.nonce.authenticate_request(ip, crypt)
         if status != 'OK':
             return status
 
         hashkey = database.nonce.random_md5()
-        prefix = hashkey[:2]
-        fullpath = '%s/full/%s' % (pngpath, prefix)
-        if not os.path.exists(fullpath):
-            os.makedirs(fullpath)
-        pngname = '%s/%s.png' % (fullpath, hashkey)
-        outfile = file(pngname, 'wb')
-        outfile.write(binary.data)
-        outfile.close()
+        save_upload(binary, hashkey)
 
-        ppmhandle, ppmname = tempfile.mkstemp()
-        error = os.system('pngtopnm "%s" > "%s"' % (pngname, ppmname))
-        assert not error
-        assert zoom(ppmname, pngpath, prefix, hashkey, 200)
-        assert zoom(ppmname, pngpath, prefix, hashkey, 400)
+        width, height, ppmhandle, ppmname = pngtoppm(hashkey)
+        if width != request_width:
+            return ("Uploaded image width (%d) is different from requested width (%d)."
+                    % (width, request_width))
+        if height > database.options.max_screenshot_height:
+            return ("Uploaded image height (%d) is greater than maximum (%d)."
+                    % (height, database.options.max_screenshot_height))
+
+        assert zoom(ppmname, hashkey, 200)
+        assert zoom(ppmname, hashkey, 400)
         os.close(ppmhandle)
         os.unlink(ppmname)
+
+        values = {'hashkey': hashkey, 'factory': factory, 'width': width, 'height': height}
+        database.insert('screenshot', values)
+        database.request.update_screenshot(request, database.lastval())
         return 'OK'
     finally:
         database.disconnect()
